refactor!: make Widget render a preact-based renderer component
BREAKING CHANGE: Heavy refactoring introduces multiple breaking changes:
- `WidgetsController.create` renamed to `add`
- Widget class internal methods have change; `refresh` is removed and
  `render` made the main rendering entrypoint
- Widget subclasses must now implement a `renderComponent` method to
  implement their rendering
