+++ src/widgets/index.ts
@@ -1,2 +1,2 @@
-export { PaymentPlanWidget as PaymentPlan } from './payment_plan'
-export { HowItWorksWidget as HowItWorks } from './how_it_works'
+// export { PaymentPlanWidget as PaymentPlan } from './payment_plan'
+export { HowItWorks } from './HowItWorks/HowItWorks'
